// src/app/Models/rankRegister.model.ts
export class rankRegisterModel {
    rankId: number;
    rankName: string;
    code: string;
    level: string;
    crewSort: string;
    groupRank: string;
    department: string;
    officeCrew: string;
    level1: string;
    modifiedBy: string;
    modifiedDate: string;
    isDeleted: boolean;
    recDate: string;
    portageSort:string;
} 